Bubble up errors rather than `panic!`ing
While `.unwrap()` on a `Result` will cause the program to `panic!` if it
fails for some reason, `try!` will simply cause the error to bubble up.
For this spike, we're still going to `panic!` at the top level, if an
error occurs at any point, but this will make more of the code easy to
use should this end up going into a real app.

Side note: The conversion between `Vec<Result<T>>` and `Result<Vec<T>>`
(possibly more, and equivalent to `sequence` in Haskell?) is simply
handled by the type system, and not something we have to write
explicitly.