Parse the skeleton
I've refactored the binary reader to have a bit less boilerplate. The
type signature is incredibly dense for the skeleton, being
`Vec<Rc<RefCell<Joint>>>`. This is because we're going to need shared
mutable aliases between the joints when we calculate the various
matrices which is always relative to the parent. Since we need this to
be bi-directional (and we certainly wouldn't want to do any clever
tree-like structure since runtime efficiency is important here, and that
includes memory locality). `Rc` allows us to have multiple references to
the value, and then `RefCell` allows us to re-introduce mutability.
Borrow checking will occur at runtime rather than compile time, and the
thread will `panic!` if more than one mutable borrow occurs at the same
time. Since there cannot be circular references between joints, this
should never be a problem.